requirements.txt
sqlmodel~=0.0.18
fastapi~=0.111.0
pydantic~=2.7.1
PyJWT~=2.8.0
g4f~=0.3.1.9
starlette~=0.37.2
passlib~=1.7.4
uvicorn~=0.30.0